[refactor]: 拆分 GamePanel.css 到多个文档中
