+++ src/Pages/Game/Board/Grid.css
@@ -1,0 +1,111 @@
+.grid {
+    --bg-color: transparent;
+    /*--stripe-color: var(--color-light3) ;*/
+
+    /*top: 0;
+    left: 0;*/
+
+    position: relative;
+    width: var(--grid-size);
+    height: var(--grid-size);
+
+    color: var(--color-light1);
+    background: var(--bg-color);
+
+    margin: var(--grid-margin);
+    padding: 0;
+
+    transition: 0.07s;
+}
+
+.grid.selectable {
+    --color-stripe: var(--color-player1);
+}
+.grid.selectable.player2 {
+    --color-stripe: var(--color-player2);
+}
+
+.grid.selectable{
+    --stripe-start: -0.5%;
+    --stripe-end: calc(var(--stripe-start) + 1.5%);
+    --stripe-gap-end: 12%;
+
+    background: repeating-linear-gradient(
+            45deg,
+            var(--color-stripe) var(--stripe-start),
+            var(--color-stripe) var(--stripe-end),
+            var(--bg-color) var(--stripe-end),
+            var(--bg-color) var(--stripe-gap-end)
+    );
+}
+.grid.selectable .piece.player2:not(.player1) {
+    background: var(--color-captureable-grid);
+}
+.grid.selectable .piece.player1:not(.player2) {
+    background: var(--color-captureable-grid);
+}
+
+.grid > div {
+    z-index: 10;
+    width: 100%;
+    height: 100%;
+
+    padding: var(--grid-padding);
+}
+.grid > div > p {
+    margin: auto calc(50% - var(--grid-font-size) / 2);
+    width: var(--grid-font-size);
+    height: auto;
+    text-align: center;
+
+    font-family: "SH Serif - SemiBold", "Source Han Serif K VF", "思源宋体 VF", "思源黑体 VF", serif;
+    font-size: var(--grid-font-size);
+    line-height: calc(var(--grid-font-size));
+    font-weight: lighter;
+
+    /*background-image: url("../Img/ShogiPiece.svg");*/
+}
+
+
+
+
+
+/* 四侧的格子去除边框 */
+.board-row:nth-child(1) .grid { /*顶部的格子*/
+    border-top: 0;
+}
+.board-row:nth-last-child(1) .grid { /*底部的格子*/
+    border-bottom: 0;
+}
+.board-row .grid:nth-last-child(1) { /*右侧的格子*/
+    border-right: 0;
+}
+.board-row .grid:nth-child(1) { /*左侧的格子*/
+    border-left: 0;
+}
+/* 四角的格子添加圆角 */
+.board-row:nth-child(1) .grid:nth-child(1),
+.board-row:nth-child(1) .grid:nth-child(1) * { /*左上角的格子*/
+    border-radius: var(--inner-border-radius) 0 0 0;
+}
+.board-row:nth-child(1) .grid:nth-last-child(1),
+.board-row:nth-child(1) .grid:nth-last-child(1) * { /*右上角的格子*/
+    border-radius: 0 var(--inner-border-radius) 0 0;
+}
+.board-row:nth-last-child(1) .grid:nth-last-child(1),
+.board-row:nth-last-child(1) .grid:nth-last-child(1) * { /*右下角的格子*/
+    border-radius: 0 0 var(--inner-border-radius) 0;
+}
+.board-row:nth-last-child(1) .grid:nth-child(1),
+.board-row:nth-last-child(1) .grid:nth-child(1) * { /*左下角的格子*/
+    border-radius: 0 0 0 var(--inner-border-radius);
+}
+
+
+/* 间隔的格子改变背景颜色 */
+.board-row:nth-child(2n) .grid:nth-child(2n+1),
+.board-row:nth-child(2n+1) .grid:nth-child(2n) {
+    --bg-color: #6665;
+    /*--stripe-color: transparent;*/
+}
+
