+++ src/Pages/Game/GamePanel.css
@@ -1,45 +1,4 @@
 
-@font-face {
-    font-family: "SH Serif - ExtraLight";
-    src: url('../../Fonts/SourceHanSerifTC-ExtraLight.otf') format('opentype');
-}
-@font-face {
-    font-family: "SH Serif - Light";
-    src: url('../../Fonts/SourceHanSerifTC-Light.otf') format("opentype");
-}
-@font-face {
-    font-family: "SH Serif - Regular";
-    src: url('../../Fonts/SourceHanSerifTC-Regular.otf') format('opentype');
-}
-@font-face {
-    font-family: "SH Serif - Medium";
-    src: url('../../Fonts/SourceHanSerifTC-Medium.otf') format('opentype');
-}
-@font-face {
-    font-family: "SH Serif - SemiBold";
-    src: url('../../Fonts/SourceHanSerifTC-SemiBold.otf') format('opentype');
-}
-@font-face {
-    font-family: "SH Serif - Bold";
-    src: url('../../Fonts/SourceHanSerifTC-Bold.otf') format('opentype');
-}
-@font-face {
-    font-family: "SH Serif - Heavy";
-    src: url('../../Fonts/SourceHanSerifTC-Heavy.otf') format('opentype');
-}
-@font-face {
-    font-family: "WHZ Regular - Bold";
-    src: url('../../Fonts/WangHanZong-Regular-Bold.ttf') format('truetype');
-}
-@font-face {
-    font-family: "WHZ SemiCursive - Medium";
-    src: url('../../Fonts/WangHanZong-SemiCursive-Medium.ttf') format('truetype');
-}
-@font-face {
-    font-family: "ZQLJ Cursive";
-    src: url('../../Fonts/ZhongQiLiuJiang-Cursive.ttf') format('truetype');
-}
-
 :root {
     --color-light1: #ddd;
     --color-light2: #999;
@@ -59,6 +18,7 @@
 
     --color-player1: #a0d5ef;
     --color-player2: #e0a5d0;
+    --color-captureable-grid: #e56;
 
     --inner-border-radius: 10px;
 }
@@ -145,73 +105,7 @@ button {
     justify-content: center;
     align-items: center;
 }
-.grid {
-    --bg-color: transparent;
-    /*--stripe-color: var(--color-light3) ;*/
 
-    /*top: 0;
-    left: 0;*/
-
-    position: relative;
-    width: var(--grid-size);
-    height: var(--grid-size);
-
-    color: var(--color-light1);
-    background: var(--bg-color);
-
-    margin: var(--grid-margin);
-    padding: 0;
-
-    transition: 0.07s;
-}
-
-.grid.selectable {
-    --color-stripe: var(--color-player1);
-}
-.grid.selectable.inverse {
-    --color-stripe: var(--color-player2);
-}
-
-.grid.selectable{
-    --stripe-start: -0.5%;
-    --stripe-end: calc(var(--stripe-start) + 1.5%);
-    --stripe-gap-end: 12%;
-
-    background: repeating-linear-gradient(
-            45deg,
-            var(--color-stripe) var(--stripe-start),
-            var(--color-stripe) var(--stripe-end),
-            var(--bg-color) var(--stripe-end),
-            var(--bg-color) var(--stripe-gap-end)
-    );
-}
-.grid.selectable .piece.inverse:not(.direct) {
-    background: #f55;
-}
-.grid.selectable .piece.direct:not(.inverse) {
-    background: #f55;
-}
-
-.grid > div {
-    z-index: 10;
-    width: 100%;
-    height: 100%;
-
-    padding: var(--grid-padding);
-}
-.grid > div > p {
-    margin: auto calc(50% - var(--grid-font-size) / 2);
-    width: var(--grid-font-size);
-    height: auto;
-    text-align: center;
-
-    font-family: "SH Serif - SemiBold", "Source Han Serif K VF", "思源宋体 VF", "思源黑体 VF", serif;
-    font-size: var(--grid-font-size);
-    line-height: calc(var(--grid-font-size));
-    font-weight: lighter;
-
-    /*background-image: url("../Imgs/ShogiPiece.svg");*/
-}
 
 .piece svg {
     position: absolute;
@@ -229,7 +123,7 @@ button {
 
 
 /*敌方棋子*/
-.piece.inverse > * {
+.piece.player2 > * {
     transform: rotate(180deg);
     transition: 0s;
 }
@@ -239,14 +133,15 @@ button {
 .piece > svg > path {
     stroke: var(--color-player1);
 }
-.piece.inverse > p {
+.piece.player2 > p {
     color: var(--color-player2);
 }
-.piece.inverse > svg > path {
+.piece.player2 > svg > path {
     stroke: var(--color-player2);
 }
 
-.grid:not(.grid:hover) {
+/* 格子边框 1px 粗细 */
+.grid {
     border: 1px solid var(--color-light2);
     border-radius: 0;
 }
@@ -254,77 +149,6 @@ button {
 .board-row {
     display: flex;
 }
-.board-row:nth-child(1) .grid {
-    /*顶部的格子*/
-    border-top: 0;
-}
-.board-row:nth-last-child(1) .grid {
-    /*底部的格子*/
-    border-bottom: 0;
-}
-.board-row .grid:nth-last-child(1) {
-    /*右侧的格子*/
-    border-right: 0;
-}
-.board-row .grid:nth-child(1) {
-    /*左侧的格子*/
-    border-left: 0;
-}
 
-.board-row:nth-child(1) .grid:nth-child(1),
-.board-row:nth-child(1) .grid:nth-child(1) * {
-    /*左上角的格子*/
-    border-radius: var(--inner-border-radius) 0 0 0;
-}
-.board-row:nth-child(1) .grid:nth-last-child(1),
-.board-row:nth-child(1) .grid:nth-last-child(1) * {
-    /*右上角的格子*/
-    border-radius: 0 var(--inner-border-radius) 0 0;
-}
-.board-row:nth-last-child(1) .grid:nth-last-child(1),
-.board-row:nth-last-child(1) .grid:nth-last-child(1) * {
-    /*右下角的格子*/
-    border-radius: 0 0 var(--inner-border-radius) 0;
-}
-.board-row:nth-last-child(1) .grid:nth-child(1),
-.board-row:nth-last-child(1) .grid:nth-child(1) * {
-    /*左下角的格子*/
-    border-radius: 0 0 0 var(--inner-border-radius);
-}
 
-.board-row:nth-child(2n) .grid:nth-child(2n+1),
-.board-row:nth-child(2n+1) .grid:nth-child(2n) {
-    --bg-color: #6665;
-    /*--stripe-color: transparent;*/
-}
 
-/*悬浮点击效果*/
-.grid:hover > div {
-    background-color: #0006;
-}
-.grid:hover::after {
-    --dp: 5px;
-
-    position: absolute;
-    top: calc(var(--dp) * -1);
-    left: calc(var(--dp) * -1);
-
-    content: "";
-    width: calc(100% + var(--dp) + var(--dp));
-    height: calc(100% + var(--dp) + var(--dp));
-    z-index: 100;
-
-    border: var(--color-light1) 3px solid;
-    border-radius: 5px;
-
-    animation: gridHover 0.1s ease-out;
-}
-
-@keyframes gridHover {
-    from { --dp: 1px; }
-    to { --dp: 4px; }
-}
-
-.grid:active {
-    background-color: #0006;
-}
